Carregar json e serializar ele, alguns ajustes nas entidades

+++ CodeChallengeGrupoZap.Domain/Entities/Location.cs
@@ -2,7 +2,7 @@ namespace CodeChallengeGrupoZap.Domain.Entities
 {
     public class Location
     {
-        public int Lon { get; set; }
-        public int Lat { get; set; }
+        public long Lon { get; set; }
+        public long Lat { get; set; }
     }
 }
